refacto: group tsc_api adapers
